server/fastapi/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.2
dataclasses-json==0.6.4
decorator==4.4.2
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.0.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.27
llvmlite==0.42.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
openai==1.14.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
orjson==3.9.15
packaging==23.2
pandas==2.2.2
pillow==10.2.0
portalocker==2.8.2
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.6.2
PyPDF2==3.0.1
pyspellchecker==0.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-pptx==0.6.23
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.17.6
waitress==3.0.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.1
wfastcgi==3.0.0
XlsxWriter==3.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.18.1